extracting qml frontend to its own file
